--- .github/workflows/deploy_prod.yml
@@ -1,44 +1,0 @@
-name: Deploy
-
-on:
-  push:
-    branches:
-      - main
-    paths:
-      - "sagemaker/**"
-      - "config.py"
-      - ".github/workflows/deploy_prod.yml"
-  workflow_dispatch:
-jobs:
-  deploy:
-    name: Deploy
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: Set up Python
-        uses: actions/setup-python@v2
-        with:
-          python-version: "3.10"
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
-
-      - name: Upload Model to S3
-        run: |
-          python -m endpoint.upload_model_s3
-        env:
-          SAGEMAKER_ROLE: ${{ secrets.SAGEMAKER_ROLE }}
-          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY }}
-          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-          AWS_DEFAULT_REGION: ${{ secrets.AWS_DEFAULT_REGION }}
-
-      - name: Create Sagemaker Endpoint
-        run: |
-          python -m endpoint.create_endpoint
-        env:
-          SAGEMAKER_ROLE: ${{ secrets.SAGEMAKER_ROLE }}
-          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY }}
-          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-          AWS_DEFAULT_REGION: ${{ secrets.AWS_DEFAULT_REGION }}
