--- .github/workflows/integration_tests.yml
@@ -1,65 +1,0 @@
-name: Integration Tests
-
-on:
-  pull_request:
-    branches:
-      - main
-
-jobs:
-  build:
-    name: Integration Tests
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: Check for relevant changes
-        id: check_changes
-        run: |
-          git fetch --depth=1 origin ${{ github.base_ref }}
-          if [ -z "$(git diff --name-only FETCH_HEAD ${{ github.sha }} -- ':(exclude).github/**' ':(exclude)docs/**')" ]; then
-            echo "No relevant changes, skipping tests."
-            echo "SKIPPING_TESTS=true" >> $GITHUB_ENV
-          else
-            echo "Relevant changes found."
-            echo "SKIPPING_TESTS=false" >> $GITHUB_ENV
-          fi
-      - name: Set up Python
-        if: env.SKIPPING_TESTS == 'false'
-        uses: actions/setup-python@v2
-        with:
-          python-version: "3.10"
-
-      - name: Cache pip dependencies
-        if: env.SKIPPING_TESTS == 'false'
-        uses: actions/cache@v2
-        with:
-          path: ~/.cache/pip
-          key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements.txt') }}
-          restore-keys: |
-            ${{ runner.os }}-pip-
-
-      - name: Cache testmon data
-        if: env.SKIPPING_TESTS == 'false'
-        uses: actions/cache@v2
-        with:
-          path: .testmondata
-          key: ${{ runner.os }}-testmon-${{ hashFiles('**/*.py') }}
-          restore-keys: |
-            ${{ runner.os }}-testmon-
-
-      - name: Install dependencies
-        if: env.SKIPPING_TESTS == 'false'
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r requirements.txt
-          pip install -r requirements-dev.txt
-          pip install -r sagemaker_model/code/requirements.txt
-
-      - name: Run tests with testmon
-        if: env.SKIPPING_TESTS == 'false'
-        run: python -m pytest -vv --testmon -m 'integration'
-        env:
-          SAGEMAKER_ROLE: ${{ secrets.SAGEMAKER_ROLE }}
-          AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY }}
-          AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
-          AWS_DEFAULT_REGION: ${{ secrets.AWS_DEFAULT_REGION }}
