+++ builder/requirements.txt
@@ -1,0 +1,14 @@
+# Required Python packages get listed here, one per line.
+# Reccomended to lock the version number to avoid unexpected changes.
+
+# You can also install packages from a git repository, e.g.:
+# git+https://github.com/runpod/runpod-python.git
+# To learn more, see https://pip.pypa.io/en/stable/reference/requirements-file-format/
+
+runpod==1.6.0
+segmentation-models-pytorch
+pytorch-lightning
+torch==2.0.1
+wandb
+rasterio
+scikit-learn
