--- pytest.ini
@@ -1,7 +1,0 @@
-[pytest]
-markers =
-    integration: marks tests as integration (deselect with '-m "not integration"')
-    e2e: marks tests as end to end (deselect with '-m "not e2e"')
-    unit: marks tests as unit (deselect with '-m "not unit"')
-    slow: marks tests as slow (deselect with '-m "not slow"')
-
