--- requirements.txt
@@ -1,3 +1,0 @@
-boto3
-sagemaker
-python-dotenv
