--- sagemaker_model/code/marinedebrisdetector_mod/__init__.py
@@ -1,1 +1,0 @@
-
