--- sagemaker_model/code/marinedebrisdetector_mod/checkpoints.py
@@ -1,29 +1,0 @@
-CHECKPOINT_FOLDERS = {
-    "unet++1": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/models/unet%2B%2B1.zip",
-    "unet++2": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/models/unet%2B%2B2.zip",
-    "unet++3": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/models/unet%2B%2B3.zip",
-    "unet++1_no_label_refinement": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/models/unet%2B%2B1_no_label_refinement.zip",
-    "unet++2_no_label_refinement": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/models/unet%2B%2B2_no_label_refinement.zip",
-    "unet++3_no_label_refinement": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/models/unet%2B%2B3_no_label_refinement.zip",
-    "unet1": (
-        "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/models/unet1.zip"
-    ),
-    "unet2": (
-        "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/models/unet2.zip"
-    ),
-    "unet3": (
-        "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/models/unet3.zip"
-    ),
-}
-
-CHECKPOINTS = {
-    "unet++1": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/checkpoints/unet%2B%2B1/epoch=54-val_loss=0.50-auroc=0.987.ckpt",
-    "unet++2": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/checkpoints/unet%2B%2B2/epoch=95-val_loss=0.65-auroc=0.988.ckpt",
-    "unet++3": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/checkpoints/unet%2B%2B3/epoch=50-val_loss=0.63-auroc=0.987.ckpt",
-    "unet++1_no_label_refinement": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/checkpoints/unet%2B%2B1_no_label_refinement/epoch=95-val_loss=0.63-auroc=0.986.ckpt",
-    "unet++2_no_label_refinement": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/checkpoints/unet%2B%2B2_no_label_refinement/epoch=82-val_loss=0.65-auroc=0.985.ckpt",
-    "unet++3_no_label_refinement": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/checkpoints/unet%2B%2B3_no_label_refinement/epoch=97-val_loss=0.64-auroc=0.990.ckpt",
-    "unet1": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/checkpoints/unet1/epoch=33-val_loss=0.59-auroc=0.985.ckpt",
-    "unet2": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/checkpoints/unet2/epoch=97-val_loss=0.63-auroc=0.982.ckpt",
-    "unet3": "https://marinedebrisdetector.s3.eu-central-1.amazonaws.com/checkpoints/unet3/epoch=73-val_loss=0.60-auroc=0.984.ckpt",
-}
