--- sagemaker_model/code/requirements.txt
@@ -1,6 +1,0 @@
-segmentation-models-pytorch
-pytorch-lightning
-torch==2.0.1
-wandb
-rasterio
-scikit-learn
