+++ src/test_input.json
@@ -1,0 +1,5 @@
+{
+    "input": {
+        "image_url": "s3://test-sagemaker-studio-768912473174-0ryazmj34j9/2400_1440.tiff"
+    }
+}
