+++ tests/data/exp_2400_1440_prediction.tiff.aux.xml
@@ -1,0 +1,11 @@
+<PAMDataset>
+  <PAMRasterBand band="1">
+    <Metadata>
+      <MDI key="STATISTICS_MAXIMUM">0.56538528203964</MDI>
+      <MDI key="STATISTICS_MEAN">0.0030319164885931</MDI>
+      <MDI key="STATISTICS_MINIMUM">7.347622510423e-12</MDI>
+      <MDI key="STATISTICS_STDDEV">0.016446156546758</MDI>
+      <MDI key="STATISTICS_VALID_PERCENT">100</MDI>
+    </Metadata>
+  </PAMRasterBand>
+</PAMDataset>
